Address review comment and split all_test.go back into multiple *_test.go files
